Updated medical scheme functionalities

+++ controller/updateClaimFormControllerOne.php
@@ -30,7 +30,7 @@
                             $_SESSION['claim_form_no'] .= "<td>{$row['claim_form_no']}</td>";
                             $_SESSION['claim_form_no'] .= "<td>{$submitted_date}</td>";
                             $_SESSION['claim_form_no'] .= "<td><a href=\"../../controller/updateClaimFormControllerTwo.php?claim_form_no={$row['claim_form_no']}&user_id={$user_id}\">Update Form</a></td>";
-                            $_SESSION['claim_form_no'] .= "<td><a href=\"../../controller/deleteClaimFormController.php?claim_form_no={$row['claim_form_no']}&user_id={$user_id}\">Delete Form</a></td>";
+                            $_SESSION['claim_form_no'] .= "<td><a href=\"../../controller/deleteClaimFormController.php?claim_form_no={$row['claim_form_no']}&user_id={$user_id}\" onclick=\"return confirm('Are you sure?');\">Delete Form</a></td>";
 
                             header('Location:../view/medicalSchemeMember/memUpdateClaimFormsV.php');
                         }
@@ -65,7 +65,7 @@
                         $_SESSION['claim_form_no'] .= "<td>{$row['claim_form_no']}</td>";
                         $_SESSION['claim_form_no'] .= "<td>{$submitted_date}</td>";
                         $_SESSION['claim_form_no'] .= "<td><a href=\"../../controller/updateClaimFormControllerTwo.php?claim_form_no={$row['claim_form_no']}&user_id={$user_id}\">Update Form</a></td>";
-                        $_SESSION['claim_form_no'] .= "<td><a href=\"../../controller/deleteClaimFormController.php?claim_form_no={$row['claim_form_no']}&user_id={$user_id}\">Delete Form</a></td>";
+                        $_SESSION['claim_form_no'] .= "<td><a href=\"../../controller/deleteClaimFormController.php?claim_form_no={$row['claim_form_no']}&user_id={$user_id}\" onclick=\"return confirm('Are you sure?');\">Delete Form</a></td>";
 
                         header('Location:../view/medicalSchemeMember/memUpdateClaimFormsV.php');
                     }
